Feat: 중복 네비게이션 개선 및 피드 이동 버튼 로직 구현
