+++ navigations/AppTabNavigator/index.js
@@ -1,8 +1,8 @@
 import CameraTab from "./CameraTab";
 import CommunityTab from "./CommunityTab";
 import FriendsTab from "./FriendsTab";
-import MainFeedTab from "./MainFeedTab";
 import MyOptionsTab from "./MyOptionsTab";
 import MyPageTab from "./MyPageTab";
 
-export { CameraTab, CommunityTab, FriendsTab, MainFeedTab, MyOptionsTab, MyPageTab };
+
+export { CameraTab, CommunityTab, FriendsTab, MyOptionsTab, MyPageTab };
