+++ navigations/Stack.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import { createStackNavigator } from "@react-navigation/stack";
-import { CommunityTab, FriendsTab, MyPageTab, MyOptionsTab, MainFeedTab, } from "./AppTabNavigator"
+import { CommunityTab, FriendsTab, MyPageTab, MyOptionsTab } from "./AppTabNavigator"
 import { View } from 'react-native';
 import { Button1, Button2, Button3 } from "../components/Button";
+import { MainFeedTabScreen } from "./AppTabNavigator/MainFeedTab";
 
 const MainFeedStack = createStackNavigator();
 
@@ -10,8 +11,8 @@ export const MainFeedStackScreen = () => { // 메인 피드 속 Stack 구성
 	return (
 		<MainFeedStack.Navigator>
 			<MainFeedStack.Screen
-				name="Logo"
-				component={MainFeedTab}
+				name="FriendFeed"
+				component={MainFeedTabScreen}
 				options={({ navigation }) => ({// 메인 피드 화면에서 상단 네비 구현
 					headerRight: () => (
 						<View style={{flexDirection: 'row'}}>
@@ -38,8 +39,8 @@ export const MainFeedStackScreen = () => { // 메인 피드 속 Stack 구성
 			/>
 			<MainFeedStack.Screen
 				name="MyOptions"
-        component={MyOptionsTab}
-      />
+				component={MyOptionsTab}
+			/>
 		</MainFeedStack.Navigator>
 	);
 }
@@ -50,7 +51,7 @@ export const CommunityStackScreen = () => { // 커뮤니티 속 Stack 구성
 	return (
 		<CommunityStack.Navigator>
 			<CommunityStack.Screen
-				name="Logo"
+				name="Community"
 				component={CommunityTab}
 				options={({ navigation }) => ({ // 커뮤니티 화면에서 상단 네비 구현
 					headerRight: () => (
